Document an error that was found
Error occurs when a path ending in a file name instead of a folder name is
specified for backup and the destination folder that is specified does not
exist.  In that case, the source file name is stored as the destination
folder name instead of creating the destination folder first.