Implement work around to avoid the error condition
In order to avoid the error documented in the previous commit, the file
backup-archive.conf is now stored in the $HOME/.config/backup folder which
becomes the backup-archive target.  All of the other single files mentioned
are in the destination folder which must be valid for the script to work.